requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
